+++ src/services/paging/index.js
@@ -1,0 +1,1 @@
+export { Pager } from './pager';